divide making appointments to component, container and logic
